Move hmac.h to public headers
