+++ src/crypto/test/bench.cpp
@@ -4,10 +4,10 @@
 #include "ccf/crypto/base64.h"
 #include "ccf/crypto/entropy.h"
 #include "ccf/crypto/hash_provider.h"
+#include "ccf/crypto/hmac.h"
 #include "ccf/crypto/key_pair.h"
 #include "ccf/crypto/sha256.h"
 #include "ccf/crypto/symmetric_key.h"
-#include "crypto/hmac.h"
 #include "crypto/openssl/base64.h"
 #include "crypto/openssl/hash.h"
 #include "crypto/openssl/key_pair.h"
